requirements.txt
albucore==0.0.17
albumentations==1.4.18
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
eval-type-backport==0.2.0
exceptiongroup==1.2.2
fire==0.7.0
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib-resources==6.4.5
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy-loader==0.4
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
narwhals==1.10.0
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.9.0
paddlepaddle==2.6.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
protobuf==5.28.2
pyarrow==17.0.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-core==2.23.4
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.39.0
tenacity==9.0.0
termcolor==2.4.0
tifffile==2023.7.10
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
zipp==3.20.2
